Fix query bug when fetching filtered entity lists
In librg_world_fetch_chunkarray and librg_world_fetch_ownerarray, only
the first buffer_limit of the world's entities were ever being checked,
even if no matching entities had yet been added to the list. This meant
that even if you know there are a small number of entites owned by a
given owner, you have to pass a big enough array to hold the whole world
to be sure to find them.